requirements.txt
# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# LLM
openai>=1.0.0
anthropic>=0.40.0

# Agent framework
langchain-core>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0

# Vector store & embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# UI
streamlit>=1.38.0

# Config
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# HTTP & resilience
httpx>=0.27.0
tenacity>=8.2.0

# PDF reports
fpdf2>=2.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
